Fix user creation issues (#3074)
userCredentials was originally not marked as required, despite being so - this has been fixed. As a result, two additional issues were discovered and fixed: NullPointerException when userCredentials is not in the payload and inserting null for userInfo in UserCredentials. The first was fixed by moving logic from the controller to the ObjectBundleHook and the second by updating UserCredentials with the correct userInfo reference before saving.

Issue: DHIS2-5732