Fix issue with missing database information in system info (#4670)
* Fixes issue with missing database information in system info (#4670)

- Database information is never cloned, but sensitive information is
reset to null on requests where user has not authority ALL. Database
information must be cloned before sensitive information is reset for
current request.

- URL is parsed to retrieve database name. Database name can be omited
with current PostgreSQL driver versions. In order to avoid further
adaptions of the parse algorithm and default value handling for upcoming
releases of the PostgreSQL JDBC driver, the resulting database name is
retrieved from database.

- URL may contain the name of the user that is used to connect to
database. To get the final user, the database can be used to return the
user name.

- Database version and URL have not been marked for output.

* Replaces clone by bean utils copy properties (#6561)