Add near-real-time analytics API support (#2975)
The main goal of this PR is to make data more quickly available in analytics tables. The overall approach is to introduce a new analytics table partition which holds only the data captured since last full analytics table update, and always include this partition in analytics queries. This can reduce the lag between analytics runs from hours down to seconds.

The PR does the following:

* Introduces a new partition table called "latest". This holds the latest data, i.e. data added or updated since last full analytics table generation. The latest partition uses a "year" value of 0.
* The latest partition is dropped and rebuilt on each run (like the regular analytics partitions).
* For each analytics table run, the system checks whether there is new, updated or deleted data since last run.
* When considering whether there is new data, the timestamp of the start of this process is used to guarantee no overlap between runs.
* Data which is updated or deleted through runs is deleted from any of the analytics partitions (including old data).
* A new synthetic unique identifier column is added to the analytics tables for the purpose of deleting updated/deleted data.
* The system checks whether new data exists since last update of the latest partition, and if not, aborts the update.

The scheduling part is yet to be done. Triggering the "latest" partition can be done through the API like this:

```
POST /api/31/resourceTables/analytics?lastYears=0&skipResourceTables=true
```