fix(dhis2-7698): use toLanguageTag instead of toString to get short code
Going through a bit of history of iso-639 reveals that the 2-char
language code for Indonesia used to be "in", but in later revisions they
changed it to "id".

Java has a policy of always being backwards compatible, so they have get
"in" intact. So if we do:

```
java.util.Locale locale = new java.util.Locale("id");
System.out.println(locale.toString());
```

It prints "in".

However, if we do this instead:

```
System.out.println(locale.toLanguageTag());
```

It prints "id", which is what we want.

Digging through
https://www.oracle.com/technetwork/java/javase/java8locales-2095355.html
we see that this is in fact the recommendation:

> Locales can be constructed with Locale.forLanguageTag(<Language Tag>) or
> Locale(<Language>, <Country>[, <Variant>]) where <Language Tag> is the
> Language Tag column value, <Language> is the ISO 639 value, <Country> is
> the ISO 3166 value, and <Variant> is the Variant column value if it’s
> neither empty nor *.
>
> Locales can be constructed only with Locale.forLanguageTag(<Language
> Tag>) if the Variant column value is *.

The star (*) denotation here is very important, as if we look up
Indonesian in the table we see: `in-ID(*)`.

So, we either have to construct our Locale with `.forLanguageTag` or,
actually convert to the language tag with `.toLanguageTag()` instead of
`.toString()`.