fix: Expire a user's active session when disabled
Before this change, when we disabled a user in the user admin, an already logged in user would still be "active/usable" until the session expired.
Now, when we disable a user via the user admin we retrieve the users's active session from the "SessionRegistry" and expire them.

Issue: [DHIS2-7358]