fix: Remove LoggerListener from the security.xml
The org.springframework.security.authentication.event.LoggerListener should not be used in production since it leaks Session ID into the logger.
The spring LoggerListener is replaced by a new custom listener, this new listener hashes the session id to prevent it from leaking into the logs.
To also be able to log auth events from http basic, the http-basic config in security.xml is augmented with a new AuthenticationDetails object that capture the x-forwarded-for header if present.

Issue: [DHIS2-7602]