fix: Add previous year to analytics partition
- DHIS2-5990
- DHIS2-8144
- When an aggregate analytics is run, certain Period Types may have the
first week of the period starting in the previous year of the given
Period Type. For instance, the Quarterly period starting on 2019/01/01
and ending in 2019/03/31 has the first week of the period starting on
2018/12/31.
Therefore, the system must take in consideration (add to the sql
where condition) the previous year, otherwise there is a mismatch when
comparing Weekly period types and Quarterly/Yearly/SixYearly and Monthly
period types.