feat: Tracker importer validation hooks
Draft check-in. This is a work in progress commit. A lot of refactoring, restructuring and testing remains.
This commit is mainly focused on getting the "CREATE" of TrackedEntities validation correct.

The test class TrackerValidationTest.java is the tests for this check-in.
The AddUpdateTrackedEntityValidationHook.java "hook" is the most important class and the main focus in this check-in.

All the tests in TrackerValidationTest should be correct...

Issue: [DHIS-7853]