fix: Ensure caching for queries with missing settings (#5161)
A performance problem is currently caused by the system making a high number of user setting queries for user settings which have no value nor default value. This causes a transaction and database lookup to happen, and consumes a fair amount of CPU because they happen frequently e.g. as part of the translation solution.

This change introduces a dedicated SerializableOptional class which is suitable for wrapping a Serializable and achieves "negative caching", which implies that null responses will be cached. This is suitable for user and system settings because there are quite few possible values. By default, caches skip null values, and so this prevents the system from re-checking missing user setting values very frequently. The SerializableOptional object is better suited as cache key than Optional as it can be type-safe.