fix: calculate metadata pagination total pages
- DHIS2-8754

This fix addresses an issue introduced with this PR: https://github.com/dhis2/dhis2-core/commit/b132f0685b9040f194ee5bb932fc08317e0a6aa8

The `Pager` object requires a "total" value in order to properly calculate the number of pages. The "total" is **now** passed to the `Pager` object and it's calculated by executing a count query using the same filter and sorting criteria used for the original query.
If the Metadata Query engine requires an in-memory query (because a filter or sort field is non-persisted), then the pagination is executed in memory by the Query Engine and no longer in the controller.