fix: prevent audit StackOverFlow error with Relationship
The Auditing component fails to serialize to Json self-referencing TEIs.
When a TEI (tei A) has a relationship with another TEI (tei B), and tei B
references tei A, the Jackson parser fails with a Stack Overflow Error.

This fix uses a Jackson annotation on the RelationshipItem class that "forces"
the serialization to only serialize the "simple" information of the linked Tei and
avoid recursion into the "relationship" attribute