feat: Period offset for Indicator formula
This feature allows to "shift" by period one or more elements within an Indicator formula.

Given the following Indicator formula:

`#{gQNFkFkObU8} + #{MaXD86iob3M}`

It is now possible to add a `periodOffset()` directive, in order to calculate the value for the
query-specific period based on the declared `periodOffset`:

`(#{gQNFkFkObU8}.periodOffset(-1) + #{MaXD86iob3M}.periodOffset(-1)) - (#{gQNFkFkObU8} + #{MaXD86iob3M})`

In the above example, assuming that the Analytics query is executed for the period Feb 2020, the results will calculate the
value of `gQNFkFkObU8` + `MaXD86iob3M` for January 2020 (offset -1) and subtract the value from February 2020.

ref: DHIS2-8758