revert: removed Tracker Audit side effect handler
The Auditing sub-system is based on Hibernate listeners that intercepts any Hibernate-managed object prior to any persistence-related operation and fire a Audit message to the Auditing sub-system.
The component that has been removed in this PR repeats the same operation from the listeners, which would cause two audit entries for each persistence operation.