fix: fix Indicator Grid results index resolution
After the introduction of Period Offset (DHIS2-8758), the Indicator data aggregation logic needs to resolve the period
index withing the aggregated Grid. In case of a query that uses ORG UNIT GROUP SET, the resulting Grid contains multiple
dimension items for each row and the period index can shift to 2 or 3 position.
With this fix, a more robust method to identify data and period index is introduced during Indicator grid handling.

ref: DHIS2-9214