fix: Remove analytics table job default value for last years [DHIS2-4664] (#5880)
Removes the default value set for AnalyticsTableUpdateParams and ContinuousAnalyticsJobParameters for the lastYears property. This should default to null and was set to 0 by a mishap. A value of 0 has special meaning and will only trigger the "latest" data.

(cherry picked from commit 5d4b8e231741221eda1e794038115d49afb09bb9)