fix: Use Org Unit uid on Events/Enrollment table creation
When a Program has a Data Element or TEA of type Organisation Unit, the system was persisting the OU name in the `analytics_event_*` and `analytics_enrollment_*` tables.

This was done to fix another issue with Event/Enrollment Line List view, which was showing the Org Unit UID rather than name.
Persisting the name, rather than the UID, was causing a bug becuase the join with the Org Unit table was broken, if a user selects "Org Unit Field" in the Event table layout (Pivot App).

This fix:
- reverts the change to the JdbcTableManager, so that the Org Unit UID is used in analytics tables
- introduces a Grid Postprocessor to resolve Grid data that have to be displayed with a name, instead of UID.
The Postprocessor has been designed to run the Grid through a number of Processors: at the moment there is only one processors in use.

ref: DHIS2-9096