fix: Use Org Unit uid on Events/Enrollment table creation (#5892)
* fix: Use Org Unit uid on Events/Enrollment table creation

When a Program has a Data Element or TEA of type Organisation Unit, the system was persisting the OU name in the analytics_event_* and analytics_enrollment_* tables.

This was done to fix another issue with Event/Enrollment Line List view, which was showing the Org Unit UID rather than name.
Persisting the name, rather than the UID, was causing a bug becuase the join with the Org Unit table was broken, if a user selects "Org Unit Field" in the Event table layout (Pivot App).

This fix:

- reverts the change to the JdbcTableManager, so that the Org Unit UID is used in analytics tables
- adds a new column "<dataelement>_uid" to the event analytics tables, containing the name of the OU
- modifies the Events query logic to use this new column when querying for events

* chore: renamed test method

(cherry picked from commit 5867c314d2db0f6c835424638d878c0f9626e8bf)