fix: [DHIS2-8588] handling division by ZERO error
Some data elements, which are not compulsory are returned as NULL values for this query. So, when converting to ZERO, we can face a division by ZERO error. This happens in cases when we have a division operation in the formula and the denominator is set to ZERO (for the cases when they are NULL).