fix: DB pool saturation under heavy concurrency
Removes the `@Transactional` annotation from the newly added `getTrackedEntityInstances2`.
The annotation was causing the C3P0 pool to saturate with 200+
concurrent users. The first 80 threads were keeping the Transaction
open (80 is the default pool size), while executing the TEI fetching
logic. The other 120 threads were also trying to access the DB, but
could not get a connection from the pool.

This PR also:

- Makes the TEI fetching use a dedicated Executor to avoid using the default JVM ThreadPool
- Adds 3 local caches for TEI Attribute Type, TEI Attribute by Program and security context
- Adds a method to persist TEI Audit in batch, rather than executing an INSERT statement for each TEI to audit
- Uses lombok annotation to remove constuctors
- Minor code cleanup