fix: [DHIS2-8588] handling division by ZERO error (#6153)
* fix: [DHIS2-8588] handling division by ZERO error

Some data elements, which are not compulsory are returned as NULL values for this query. So, when converting to ZERO, we can face a division by ZERO error. This happens in cases when we have a division operation in the formula and the denominator is set to ZERO (for the cases when they are NULL).

* fix: [DHIS2-8588] Code formatting

* fix: [DHIS2-8588] Code formatting

* fix: [DHIS2-8588] Sonarqube issue

* Update dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/event/data/JdbcEventAnalyticsManager.java

Co-authored-by: Stian Sandvold <stian@dhis2.org>