perf: use static cache in AbstractController
Spring will create many instances of `AbstractController`, since each Metadata controller is extending that abstract class.
Since the cache2k cache in `AbstractController` is not declared as static, we end up with many caches which are completely wasteful.
By adding the `static` keyword we make sure that one cache is shared among controllers.