Group table changes into one statement.
On large tables, running the alter column multiples times in a row
will be very slow and will result in a lot of disk space being
needed.

By running this once, postgres will be able to alter the columns in
one go, this reduces the time considerably.

I've ran this command before doing the migration, the datavalue ones
took: 30 minutes, the datavalueadit took 46 minutes. But once those
were done, running the migration took only 15 minutes.

When I tried to run the migration as is, it would take multiple hours
and be very disk space hungry.