fix: allow validation to reference entities in the payload
This PR resolves an issue where a payload having NEW entities referencing each other (eg. an Enrollment referencing a Tracked Entity Instance)
made the validation process to fail.
The validation process was only referencing "pre-heated" entities - that is, entities that were loaded from the database and placed into the pre-heat cache.
In case of new entities, the validation code was not able to resolve the references to new entities, as it was trying to find them in the pre-heat cache.
This PR:
- Introduces a new tree-based data structure in the preheat that "tracks" payload-only entities.
- Changes the validation logic to reference the new structure when checking for object references.
Note that this PR is based on an older PR, which contains a refactoring of the pre-heat code: https://github.com/dhis2/dhis2-core/pull/6560
ref: DHIS2-9288