fix: incorrect ignore count during Tracker Import
This PR fixes the value of the `ignore` field of the Tracker Import Report
when Atomic Mode is set to `ALL`.

Additionally, this PR contains a refactor of the `TrackerBundleService`: the persistence code has been split into dedicated sub-classes and the persistence flow has been refactored into a template method, in order to avoid duplication and simplify the logic.
A new unchecked exception is thrown from the persistence flow (on AtomicMode ALL) to roll-back the transaction, in case of an error during persistence.