fix: remove long running transaction during Tracker Import
This PR introduces a "mapping layer" in the Tracker Import pre-heater, in order to detach the pre-heated objects from the Hibernate Session
and prevent `LazyInitializationException` errors during the validation and processing stage of the Import.
Thanks to this new layer, it is possible to remove the `@Transactional` annotation from the Tracker Service `import` method and open fast-running transactionz only during the preheat stage and the persistence stage.

The role of the mapping layer is to map the Hibernate-managed objects fetched by the preheater to the same object. This operation has two benefits:

- initialize `lazy` collections which are required during the validation stage and would normally fail, without a Transaction
- reduce the amount of data stored in the pre-heat layer

The new mapping layer is based on the [MapStruct](https://mapstruct.org/) library.
This library is very efficient because it does not use reflection: instead it generates the mapping code at build time.
The current overhead of the mapping layer is negligible.
For a more detailed description of this layer, please see the [SAD document](https://docs.google.com/document/d/1VgFhJecVipoQps6Th8VZbMnbn6YqfyLWi93_SrMUhBQ/edit?usp=sharing).

Additionaly this PR:

- removes `TrackerBundleParams` and `PreheatParams`, simplifying the propagation of import arguments across the layers
- simplifies the propagation of the `User` object across the layers
- refactors the Tracker Import unit and integration test cases for better long term maintenance and clarity
- moves the `TrackerBundleParamsConverter` to the controller layer