fix: Tracker Identifiers resolution during Tracker import
This PR simplifies the way the `TrackerPreheat` object keeps track of metadata objects during the Tracker Import.
Rather than keeping track of the `TrackerIdScheme` associated to each class of object, the system now simply uses a `Map` where
the key is the class type (e.g. `Program`) and the value is a `Map` of object key and pre-heated object.

Assuming that IdScheme `CODE` is used for `Program`, the data is structred like so:

```
Program.class -->  CODE_1 -> Program 1
                   CODE_2 -> Program 2
                   CODE_3 -> Program 3

```

By using this simplified data structred, the pre-heated object can be simply fetched by using the id used in the payload, without having to specify the `TrackerIdScheme`.

Additionally, this PR fixes a serialization issue with the `TrackerIdentifierParams`, which was not serialized correctly to Json and
"broke" the propagation of IdScheme to the Import Service.

ref: DHIS2-9895