fix: DHIS2-10082: Fix daily reporting rate calculation (#6835)
* fix:DHIS2-10082: Fix daily reporting rate calculation

Currently, when we have the Periods defined as filter and the DataSet has period type set to Daily, the calculation does not work. This change will consider multiple daily Periods as filter and fix this issue. (2.34)

* fix: DHIS2-10082: Fix daily reporting rate calculation

Currently, when we have the Periods defined as filter and the DataSet has period type set to Daily, the calculation does not work. This change will consider multiple daily Periods as filter and fix this issue. (2.34)