fix: add react to dhis-web-commons
After removing core-resource app in 2.35 some implicit dependencies to
it has remained. For example the d2-ui header bar assumes React to be
available, and now that it is not, we need to ensure that it will be.

It needs to be loaded with the other global JavaScript libraries, such
as jQuery.

This is temporary and will go away when the Data Entry app is ported to
a standalone web app.

Jira-ticket: DHIS2-10097