perf: replace expensive "Program has OrgUnit" call with faster SQL
Deprecated the method `Program.hasOrganisationUnit()` and replaced with SQL method on the `ProgramService`.

All calls to `Program.hasOrganisationUnit()` have been replaced with a call to `ProgramService.hasOrgUnit`, which uses a fast JDBC-based query against the `program_organisationunits` table.

Additionally, added a check to disable the `LOAD` Hibernate listener, if ON_LOAD auditing is disabled by the user.