fix: Set default value to true for empty database after test method  (#6913)
Currently DhisTest and IntegrationTestBase have an abstract method emptyDatabaseAfterTest. Now all concrete test classes return true for this method, which means it is more feasible to have true as the default value and then allow classes to override this method in the rare case it should be set to false.