perf: reduce number of Org Units fetched from DB on Tracker Import
This PR modifies the Event import code that was refactored in version 2.35 and optimmizes the check to verify if an Event's Org Unit is part of the Event's Program Org Units List.
Programs are now stored in the `WorkContext` without the "link" to Organization Units.
A new, dedicated data structure is created to map Program to Org Units - using only Org Units effectively specified in the payload.

This change is required to minimize memory allocation and SQL exec time when a Program has a very large (> 10.000) Org Units associated to it.

ref: DHIS2-10139