fix: send message rule action when stage is part of registration
JIRA #DHIS2-9935
PR fixes the problem with notification events. The event will be invoked in case of insert/update, previously it was only being done for updates.
The change includes the addition of fallbackExecution = true flag in @TransactionalEventListener annotation to make sure the event is invoked even if there is no transaction running.