feat: Config property for disabling server-side program rules [DHIS2-10380] (#7227)
Adds a new DHIS 2 config property system.program_rule.server_execution = on | off. Default value is on and there is no change in current behavior. When facing high CPU usage in production, server-side program rules are sometimes the main cause. From a sysadmin perspective it is useful to be able to quickly disable server-side program rules without doing invasive changes to the metadata.