feat: check data integrity using DB queries [DHIS2-10408] (#7341)
The goal of this PR is to substitute methods in the DataIntegrityService that would fetch all rows from DB and do in-memory filtering and aggregation with equivalent implementations that would do the bulk or all of the work in the DB already.
The main reason is to keep memory (and CPU) usage reasonable in case there are many rows in the database. In particular in connection with organisation units.