test: data integrity integration tests [DHIS2-10408] (#7711)
This PR adds a new spring controller integration test DataIntegrityControllerTest that uses the REST API to setup and check the data integrity report. This is done as a part of and in agreement with QA since the scenarios tested presumably cannot be setup using the UI. Most can be setup using REST API but require knowledge of the API details.