feat: followup analysis with UIDs [DHIS2-10658] (#7906)
Makes the inputs and outputs of follow up analysis (which is a read-only request) accessible to a world of UIDs (instead of exposing and requiring internal numeric database table IDs).

Has all the features of the replaced ID version but also accepts more combinations of inputs (see javadoc) and does sharing access checks for the DataElements. Also everything happens in the database instead of doing parts in the object model.

The request is validated to make sure all required parameters are present and reasonable.