fix: Require specific authority to generate min-max values [DHIS2-11157] (#8024)
This patch requires the existing F_GENERATE_MIN_MAX_VALUES authority for generating min-max values through the min-max value API endpoint. This action required the F_PERFORM_MAINTENANCE authority. It is arguable better to require the more specific authority as it actually exists in the system (this was not carried over properly from the old Struts web module).