fix: Non-optimized TEI fetch for events access control [DHIS2-11264]
This fix removes the fetch of TEI per event as part of the get events method for with-registration programs, by moving the check for whether the program is a without-registration program before the TEI fetch. The TEI fetch is redundant in the case of without-registration program as the program type is known up front.