perf: Remove program lookup in event query loop [DHIS2-11264] (#8149)
This patch is a performance improvement for the event query API. With this fix, instead of looking up the program from the database for every event in a loop and subsequently using the program type to determine whether the access check can be skipped, the program type is returned from the event query as part of the event model. This removes the single-database-in-loop anti-pattern.

(cherry picked from commit 15a1ce87df3a6aecbb40ceca8c1aeb9ef9a35290)