feat: Org unit merge [DHIS2-11281] (#8173)
This issue is about introducing support for merging organisation units with related metadata, aggregate data and tracker data. A typical scenario is two or more administrative districts which have merged into one district. The solution uses a mix of Hibernate/JPA, JPQL and SQL to merge various domain entities. For testing merge handlers using JPQL or SQL, the test code also uses JPQL or SQL to make sure the changes are visible to the current transaction, as JPQL update statements are not visible to the Hibernate session within Spring-managed tests wrapped in a transaction.