fix: Installed apps disappear when caches are cleared [DHIS2-11451]
When application caches are cleared as part of the maintenance operation, the process also clears the app cache. This cache is used as the primary storage for installed apps and is not meant to be cleared during normal runtime. Moving this cache out of the cache provider to avoid it being cleared by the maintenance app clear operation.