chore: use one Dockerfile
We do not need the extract logic with an intermediary slim docker image
as a placeholder for the war. We can leverage docker multi-stages and
copy from the build stage into any other stage.

It will build the debian image by default and with --target
like so

docker build . -t dhis/core:alpine --target alpine

build the alpine image