fix: NTI: time should be ignored when validating enrollment date DHIS2-11973 (#9039)
* chore: remove unused method

* fix: support parsing Instant for tracker reports

* chore: do not expose TrackerReportUtils

Util classes should be avoided IMHO as they become the dumping ground for
what we do not know where else to put. Reduce its visibilty to make this
harder. If time permits, think about a different name or place for it

* fix: todays enrollment dates are not in the future DHIS2-11973

requests from the UI only send in a date like "2021-10-12" but we get
API requests with a time portion like "2021-10-12T14:00:00.000"

only consider the date portion of the enrollment date & incident date
when validating programs which do not allow future dates