chore: reduce usage of commons-lang TECH-777
Some module depended on commons-lang to be a transitive dependency from
dhis-api. Thats why I added some direct dependencies on commons-lang3
for them.

StringEscapeUtils.escapeSql will need to be replaced by another library
since its removed from commons-lang3
https://commons.apache.org/proper/commons-lang/article3_0.html

apache velocity still needs common-lang. It was erroneously
excluded as it is cleary needed by velocity. The tests show that
velocity fails to work at runtime if this library is missing. So we
cannot just yet remove it from the war file. Velocity itself has to be
upgraded. Same goes for jasperreports that is a pretty old version that
needs commons-lang.