chore: reduce docker build context (#9038)
* chore: reduce docker build context

every time we run 'docker build' the docker client archives all files in
the build context and sends them to the daemon. for me these were ~
4.63GB which probably depends on whether you ran a local maven build
before or not.

this change reduces docker build context and thus speeds up the docker
build start time but also the entire docker build as 'COPY . .' now only
copies what is actually needed for the build. this also impacts docker
cache invalidation as previously any file that was added/changed would
invalidate the 'COPY . .' layer. That means files irrelevant to the
build were causing the maven build to be repeated ~ 7 minutes for me.

another aspect that might not be relevant in this scenario (as its only
concerning the build stage) is security. secure tokens stored in
gitignored files in this repo can be included in a docker image if not
excluded from the docker build context.

Here are some stats with before/after timings/sizes. docker builds were
run with buildkit enabled 'DOCKER_BUILDKIT=1 docker build .'

docker build context load time/size
before:
 => [internal] load build context 20.2s
 => => transferring context: 4.54GB
after:
 => [internal] load build context 0.7s
 => => transferring context: 55.43MB 0.7s

COPY . .
before: 9.1s
after: 0.6s

Note that the timings may vary for you and also on every run.

For more read https://codefresh.io/docker-tutorial/not-ignore-dockerignore-2/

* chore: build image does not need latest package updates

the build image is temporary, thus not used in production. the build
tools themselves are kept up to date with the latest official maven
build image we use as the base. we can save ourselves the occasional
~ 15 seconds in case this layer would need to be run.