chore: temporarily archive integration test surefire reports TECH-784 (#9149)
to investigate how integration test runs that take 22min differ from the ones that run 15min
Surefire gives detailed reports in XML that can automatically be parsed and analyzed

I set the retention days to less than the default of 90 days to not eat
up to much of our quota. We can also remove this once the investigation
is over.