chore: build but exclude analytics tests from integration-test TECH-784 (#9216)
* chore: build but exclude analytics tests from integration-test TECH-784

excluding the entire project from the maven build will cause a
dependency of the report service to be missing. So we have to build it
but exclude its tests from executing.

The same goes for the analytics module. We need to build its
dependencies and than only run its tests.

* chore: exctract building analytics with deps into own step TECH-784

this shows us more clearly how long it takes to build & run