fix: explicitly declare direct dependencies in dhis-api (#9066)
* fix: declare dependency on org.springframework:spring-tx in dhis-api

TrackedEntityAttributeService depends on it

but we are not explicitly declaring this dependency. Its best to be
explicit about the dependencies. This ensures we are in charge of their
versions and certain they will be there. Transitive dependencies can
change.

* fix: declare dependency on org.springframework:spring-core in dhis-api

AppManager and more dependend on it

but we are not explicitly declaring this dependency. Its best to be
explicit about the dependencies. This ensures we are in charge of their
versions and certain they will be there. Transitive dependencies can
change.

* fix: declare dependency on jackson-databind in dhis-api

~180 imports in dhis-api

but we are not explicitly declaring this dependency. Its best to be
explicit about the dependencies. This ensures we are in charge of their
versions and certain they will be there. Transitive dependencies can
change.

* fix: declare dependency on org.locationtech.jts:jts-core in dhis-api

like CoordinateObject, OrganisationUnit and more

but we are not explicitly declaring this dependency. Its best to be
explicit about the dependencies. This ensures we are in charge of their
versions and certain they will be there. Transitive dependencies can
change.