chore: schedule dependabot outside 'normal' working hours
there is a limit of 20 concurrent jobs (I think for the entire
organization) that we can run
https://docs.github.com/en/actions/learn-github-actions/usage-limits-billing-and-administration\#usage-limits

Run dependabot updates at times where most are probably not working, so
it does not take up the quota. dependabot would otherwise randomly
choose a time which can fall into our working day.

Choosing an odd time as GitHub says that high load times are at the
start of every hour.
https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows\#scheduled-events