chore: reduce total open dependabot PRs
Let's take it a bit slower with creating PRs and focus on testing fewer
more thorougly before merging them in. Also any change to the
depenabot.yml can trigger PRs to be opened outside of the set schedule
interval. So having fewer also reduces the likelyhood of it taking up
our quota of 20 total concurrent jobs
https://docs.github.com/en/actions/learn-github-actions/usage-limits-billing-and-administration\#usage-limits