chore: quiet down Jenkins and local build
We already did this in https://github.com/dhis2/dhis2-core/pull/9207 for most GitHub actions.
maven prints progress about downloading dependencies which we do not need to see. It just clutters,
increases the logs which makes finding things that one is looking for hard (errors, warnings, ...)

This should create cleaner build logs locally when using build-dev.sh,
when looking at the api test logs (which we run on GitHub) and when
looking at logs on Jenkins.