chore: provide meaningful error message DHIS2-12123
the failed event import in NTI is due to a missing reference from a
program stage to a program. this should not happen but it does
occasionally, due to the way the metadata import is implemented.

provide an error message guiding the user to a solution for this edge
case, until the root cause is fixed.