fix: fix running dhis war using maven jetty plugin
https://www.eclipse.org/jetty/javadoc/jetty-9/org/eclipse/jetty/server/NCSARequestLog.html
was deprecated and subsequently removed

Jetty 10 changed quite a lot regarding logging
https://github.com/eclipse/jetty.project/issues/4572

The Jetty libraries (both client and server) use SLF4J as logging APIs.
The only config we had in jetty-logging.properties is thus not needed
anymore.

There might be other settings we find have changed, that we need to adapt