chore: provide meaningful error message DHIS2-12123 (#9312)
* chore: provide meaningful error message DHIS2-12123

the failed event import in NTI is due to a missing reference from a
program stage to a program. this should not happen but it does
occasionally, due to the way the metadata import is implemented.

provide an error message guiding the user to a solution for this edge
case, until the root cause is fixed.

* fix: add programs without registration to preheat DHIS2-12123

https://docs.dhis2.org/en/develop/using-the-api/dhis-core-version-master/new-tracker.html\#events
states that program is not required. In this case the program and then
program instance are not in the preheat. This leads to an error in
the SQL insert statement as program instance needs to be specified
on the program stage instance. Since programs without registration
are expected/estimated to be in the order of a dozen we are fetching
all of them