fix: validate program stage reference to program DHIS2-12123
instead of throwing a more detailed exception.

* EventProgramPreProcessor
* EventWithoutRegistrationPreProcessor
are 2 pre-processors where we discovered the NPE due to a missing
reference from program stage to program. This is an edge case due to how
the metadata import is currently implemented.

Since pre-processing is done before validation we opted for returning
before an NPE occurs. Validation will then find that a reference to a
program is missing and create an error report.