fix: validate program stage reference to program DHIS2-12123
instead of throwing a more detailed exception.

* EventProgramPreProcessor
* EventWithoutRegistrationPreProcessor

are 2 pre-processors where we discovered an NPE due to a missing
reference in the DB from program stage to program. This is an edge case
due to how the metadata import is currently implemented.

Since pre-processing is done before validation we opted for returning
before an NPE occurs. Validation will then find that a reference to a
program is missing in the DB and create an error report.

Note that the program field is not a mandatory field in the payload
according to our NTI API documentation. I therefore needed an extra
error code and return before adding E1123 which would only confuse the
user.