chore: return to not depending on EasyRandom
modules depended on the support-test BeanRandomizer for creating the random beans.
Return to this style as it decouples us from BeanRandomizer. This can easily be changed
if we need to. Also individual tests with specific needs for configuring the EasyRandom
using its EasyRandomParameters can do so by opting out of BeanRandomizer