fix: maven dependency:analyze declares lombok as unused
we declare lombok as scope provided since its a compile time dependency that does not need
to be present at runtime. maven dependency:analyze often tells us lombok is declared but
unused because it does its work at compile time, and as such most annotations are removed after compilation.

We can tell the dependency plugin to ignore such 'declared unused' dependencies since with scope
provided it will not be shipped in our final war even if it was actually unused in one of our modules

Same with the hamcrest-library which is needed until we update to junit
5 which does not depend on hamcrest anymore. See http://hamcrest.org/JavaHamcrest/distributables#maven-upgrade-example