chore: hooks are always enabled
remove isEnabled to simplify code as hooks are always enabled. the
validation code is pretty complex with order playing a role and multiple
abstract classes and template methods. Remove unnecessary code as much
as possible to enable more profound refactorings. If you do not want a
hook to be enabled remove it.