chore: simplify tracker import validation TECH-880 (#9510)
* chore: add test to ensure ValidationConfig injection and validation order

* chore: remove redundant instantiation

the ValidationSerice already returns a new TrackerValidationReport

* chore: hooks are always enabled

remove isEnabled to simplify code as hooks are always enabled. the
validation code is pretty complex with order playing a role and multiple
abstract classes and template methods. Remove unnecessary code as much
as possible to enable more profound refactorings. If you do not want a
hook to be enabled remove it.

* chore: remove ability to influence order within a hook

depending on order has a few problems:
* adds complexity (more code)
* makes the validation harder to test
* prevents running validations concurrently as they depend on each other
* harder to reason about

* chore: remove hook ordering code

The order of validation hooks was the declaration order in the
TrackerValidationConfig beans. So the code was just ceremony to achieve
what was already achieved in the declaration. Inject beans directly into
the DefaultTrackerValidationService using the pattern we use in the
dxf2 checkers where order matters as well.

* chore: pass mocks to ValidationService without reflection

* chore: address comments