chore: rely on one ValidationErrorReporter
We previously created one ValidationErrorReporter per hook, which then
needed to be merged into the "root" reporter. I think this was only
necessary because of the TrackerValidationHook signature which returned
ValidationErrorReporter instead of passing one in.

Adapting the TrackerValidationHook signature to address that, leads to a
single report being created in the DefaultTrackerValidationService,
passed to hooks which add errors. No merging needed. The flow is now
easier to follow. This also fixes a minor issue that an invalid DTO is
not removed from the ValidationErrorReporter.invalidDTOs in FAIL_FAST
mode. That was because merging was skipped in that case because of the
ValidationFailFastException being thrown. Now invalidDTOs are removed
immediately on addError().