chore: inline ImportAccessManager
it is only used in the PreCheckSecurityOwnershipValidationHook. It was not tested,
only mocked in the PreCheckSecurityOwnershipValidationHookTest which could hide
ACL issues. It also obscures what is happening. The checks are not complicated
but plenty. Now its clear that the Hook throws a lot of exceptions instead of
creating error reports. Inlining will also make it easier to refactor
the ValidationErrorReporter and how hooks add errors to it => remove merge
logic in the ValidationErrorReporter